requirements.txt
absl-py>=0.7.0
numpy>=1.13.3
autograd
scipy
